+++ lib/ArduinoJson/src/ArduinoJson/JsonObjectImpl.hpp
@@ -14,15 +14,19 @@ template <typename TStringRef>
 inline JsonArray &JsonObject::createNestedArray_impl(TStringRef key) {
   if (!_buffer) return JsonArray::invalid();
   JsonArray &array = _buffer->createArray();
-  set(key, array);
-  return array;
+  if (set(key, array))
+    return array;
+  else
+    return JsonArray::invalid();
 }
 
 template <typename TStringRef>
 inline JsonObject &JsonObject::createNestedObject_impl(TStringRef key) {
   if (!_buffer) return JsonObject::invalid();
   JsonObject &object = _buffer->createObject();
-  set(key, object);
-  return object;
+  if (set(key, object))
+    return object;
+  else
+    return JsonObject::invalid();
 }
 }  // namespace ArduinoJson