+++ src/struct.h
@@ -1,0 +1,41 @@
+#ifndef STRUCT_H
+#define STRUCT_H
+
+enum{
+    MODE_TOUR_1 = 1,
+    MODE_TOUR_2,
+    MODE_TOUR_3,
+    MODE_TOUR_4,
+    MODE_PRESET,
+    MODE_UNKNOW
+};
+
+enum{
+    SERVO_1 = 0,
+    SERVO_2,
+    SERVO_3,
+    SERVO_4,
+    SERVO_5,
+    SERVO_6,
+    SERVO_7,
+    SERVO_8,
+    SERVO_9,
+    SERVO_10,
+    SERVO_11,
+    SERVO_12,
+    SERVO_13,
+    SERVO_14,
+    SERVO_15,
+    SERVO_16,
+    SERVO_17,
+    SERVO_18,
+    SERVO_19,
+    SERVO_20
+};
+
+enum{
+    SERVO_STOP = 0,
+    SERVO_RUNNING
+};
+
+#endif  //STRUCT_H